ml-face-service/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
numpy>=1.24.0
Pillow>=10.0.0
face-recognition>=1.3.0
cmake>=3.27.0
dlib>=19.24.0
